Fix C bad practices: use intptr_t for hash tables, improve error handling
- Replace long with intptr_t in hash table API for 64-bit pointer safety
- Add NULL checks for malloc in initHashTable and loadFile
- Fix open() return check (== -1 instead of !fd) and close fd on fdopen failure
- Replace unsafe strcpy/strcat with snprintf in vl_expandPath and getEnvironVars
- Make loadFile return different error codes (-1 to -4) and update all callers
  to check < 0 instead of == -1
- Initialize *strp = NULL at start of loadFile to prevent garbage pointers
- Make makeAbsolutePath always return newly allocated string for consistent
  memory ownership
- Fix passthru_argument memory leak: initialize to NULL, free before reassign
- Use uintptr_t for modulo operations to avoid negative key issues
- Add size validation in initHashTable to prevent division by zero
- Add cmp_int_stg wrapper for libstg API compatibility

+++ src/viola/mystrings.c
@@ -29,14 +29,14 @@ int numOfGBuffs = 0;
 char* GBuff[NUM_OF_GBUFFS];
 int GBuffIdx[NUM_OF_GBUFFS];
 
-int cmp_int(int n1, int n2)
+intptr_t cmp_int(intptr_t n1, intptr_t n2)
 {
     return n1 == n2;
 }
 
-long cmp_str(long a, long b)
+intptr_t cmp_str(intptr_t a, intptr_t b)
 {
-    return (long)strcmp((char*)a, (char*)b);
+    return (intptr_t)strcmp((char*)a, (char*)b);
 }
 
 void SkipBlanks(char* linep, int* i)
